+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes to Calva.
 
 ## [Unreleased]
 
+- [Use latest `cljfmt` (`0.11.2`)](https://github.com/BetterThanTomorrow/calva/issues/2274)
+
 ## [2.0.382] - 2023-07-30
 
 - Fix: [Calva can't seem to parse a ns form with metadata after 2.0.376](https://github.com/BetterThanTomorrow/calva/issues/2266)